+++ store/store-connection-manager/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Minor Changes
 
-- 03cdccc6e: New option added: disableRelinkFromStore.
+- 03cdccc6e: New option added: disableRelinkLocalDirDeps.
 
 ### Patch Changes